backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29.0",
    "cohere>=5.20.0",
    "fastapi>=0.124.0",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "openai-agents>=0.6.2",
    "openai-chatkit>=1.4.0",
    "passlib[argon2]>=1.7.4",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.16.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "tenacity>=9.1.2",
    "trafilatura>=2.0.0",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "models", "scripts"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
